c++17

+++ src/cpp/src/visual_language/inputs_embedder.cpp
@@ -402,8 +402,8 @@ private:
                 for (size_t d = 0; d < embed_dim / 2; ++d) {
                     // Correctly access the 2D position grid
                     float value = omega[d] * pos_data[h * W + w];
-                    emb_data[h * W * embed_dim + w * embed_dim + d] = std::sinf(float(value));
-                    emb_data[h * W * embed_dim + w * embed_dim + d + (embed_dim / 2)] = std::cosf(float(value));
+                    emb_data[h * W * embed_dim + w * embed_dim + d] = std::sin(value);
+                    emb_data[h * W * embed_dim + w * embed_dim + d + (embed_dim / 2)] = std::cos(value);
                 }
             }
         }
